### 2025-03-12 #692 @kelliejarcher
hdcuremodels
Submitting Author Name: Kellie J. Archer
Submitting Author Github Handle: @kelliejarcher
Repository:  https://github.com/kelliejarcher/hdcuremodels
Version submitted: 0.0.2
Submission type: Stats
Badge grade: silver

Version accepted: TBD

---

```
Package: hdcuremodels
Title: Penalized Mixture Cure Models for High-Dimensional Data
Version: 0.0.2
    c(person("Han", "Fu", role = "aut"), person(c("Kellie J."), "Archer", email=
    "archer.43@osu.edu", role = c("aut","cre"), comment = c(ORCID="0000-0003-1555-5781")))
Description: Provides functions for fitting various penalized parametric and semi-parametric mixture cure models with different penalty functions, testing for a significant cure fraction, and testing for sufficient follow-up as described in Fu et al (2022)<doi:10.1002/sim.9513> and Archer et al (2024)<doi:10.1186/s13045-024-01553-6>. False discovery rate controlled variable selection is provided using model-X knock-offs.
Imports: doParallel,
         flexsurv,
         flexsurvcure,
         foreach,
         ggpubr,
         glmnet,
         knockoff,
         mvnfast,
         parallel,
         plyr,
         survival
Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
    roxygen2,
```

- Please indicate which of our [statistical package categories](https://stats-devguide.ropensci.org/overview.html#overview-categories) this package falls under. (Please check one or more appropriate boxes below):

- [x] A pre-submission inquiry has been approved in [issue#690](#<issue_num>)

Analysts who model time-to-event outcomes when some subjects either experience long-term survival or are not susceptible to the event of interest (simplistically, cured).

-   Paste your responses to our [*General Standard* **G1.1** *The first implementation of a novel algorithm*](https://stats-devguide.ropensci.org/standards.html#general-standards), describing whether your software is:

    - *The first implementation of a novel algorithm*; or
    - *The first implementation within* **R** *of an algorithm which has previously been implemented in other languages or contexts*; or
    - *An improvement on other implementations of similar algorithms in* **R**.

    Please include hyperlinked references to all other relevant software.

-    What grade of badge are you aiming for? ([silver](https://stats-devguide.ropensci.org/pkgdev.html#pkgdev-badges))
Silver

Have a demonstrated generality of usage beyond one single envisioned use case.

- [x] Do you intend for this package to go on CRAN? I submitted 0.0.1 version of hdcuremodels last June and then learned about ROpenSci. I will not submit a new version to CRAN until after the ROpenSci review.
