### 2025-03-20 #692 @maurolepore
@kelliejarcher thanks so much for following up with a full submission.

Here I share some preliminary checks to support conversations with potential handling editors. If any of this feedback feels useful, now it's a great time to incorporate it.

We'll come back ASAP.

### EiC pre-checks

        > Lacks a pkgdown website.
        > The file names in R/ suggest an alpha sorted reference might not group functionality in a meaningful way.
        > Consider renaming the file to match the  package name so that pkgdown automatically adds a "Get started" tab on the website.

    ```
    mv vignettes/hdcuremodels-vignette.Rmd vignettes/hdcuremodels.Rmd
    ```

    > Consider adding instructions to install the development version
    > I see `set.seed()` and no cleanup. Consider using `withr::local_seed()` for auto-cleanup.
    > I see multiple calls to `expect_error()` with no string, class, or snapshot to match. Consider making the expectation more specific.
