### 2025-04-22 #692 @kelliejarcher
This is my first ROpenSci submission. I previously made some revisions based on comments below. Can you please explain what the next steps will be?

Thanks for any information,
Kellie

From: Mauro Lepore ***@***.***>
Date: Thursday, March 20, 2025 at 9:37 AM
To: ropensci/software-review ***@***.***>
Cc: Archer, Kellie ***@***.***>, Mention ***@***.***>
Subject: Re: [ropensci/software-review] hdcuremodels (Issue #692)
@ kelliejarcher thanks so much for following up with a full submission. Here I share some preliminary checks to support conversations with potential handling editors. If any of this feedback feels useful, now it's a great time to incorporate

@kelliejarcher<https://urldefense.com/v3/__https:/github.com/kelliejarcher__;!!KGKeukY!wMTZYbC16UJ1v3pEX3gwH3zgylG89i2yJNoS7MGjJjcG1YV6ApiskzQOwrP6Xs1icqztS1NNQMmA8-jDXGacZeF_ug$> thanks so much for following up with a full submission.

Here I share some preliminary checks to support conversations with potential handling editors. If any of this feedback feels useful, now it's a great time to incorporate it.

We'll come back ASAP.

EiC pre-checks

  *   [] Documentation: The package has sufficient documentation available online (README, pkgdown docs) to allow for an assessment of functionality and scope without installing the package. In particular,

     *   Is the case for the package well made?
     *   Is the reference index page clear (grouped by topic if necessary)?

Lacks a pkgdown website.
The file names in R/ suggest an alpha sorted reference might not group functionality in a meaningful way.

     *   Are vignettes readable, sufficiently detailed and not just perfunctory?

Consider renaming the file to match the package name so that pkgdown automatically adds a "Get started" tab on the website.

·         mv vignettes/hdcuremodels-vignette.Rmd vignettes/hdcuremodels.Rmd

  *   Fit: The package meets criteria for fit<https://urldefense.com/v3/__https:/devguide.ropensci.org/policies.html*package-categories__;Iw!!KGKeukY!wMTZYbC16UJ1v3pEX3gwH3zgylG89i2yJNoS7MGjJjcG1YV6ApiskzQOwrP6Xs1icqztS1NNQMmA8-jDXGaFbCIAYg$> and overlap<https://urldefense.com/v3/__https:/devguide.ropensci.org/policies.html*overlap__;Iw!!KGKeukY!wMTZYbC16UJ1v3pEX3gwH3zgylG89i2yJNoS7MGjJjcG1YV6ApiskzQOwrP6Xs1icqztS1NNQMmA8-jDXGZ9ovjSRw$>.
  *   Installation instructions: Are installation instructions clear enough for human users?

Consider adding instructions to install the development version

  *   Tests: If the package has some interactivity / HTTP / plot production etc. are the tests using state-of-the-art tooling<https://urldefense.com/v3/__https:/devguide.ropensci.org/building.html*testing__;Iw!!KGKeukY!wMTZYbC16UJ1v3pEX3gwH3zgylG89i2yJNoS7MGjJjcG1YV6ApiskzQOwrP6Xs1icqztS1NNQMmA8-jDXGZFrSq6Fw$>?

I see set.seed() and no cleanup. Consider using withr::local_seed() for auto-cleanup.
I see multiple calls to expect_error() with no string, class, or snapshot to match. Consider making the expectation more specific.

  *   Contributing information: Is the documentation for contribution clear enough e.g. tokens for tests, playgrounds?
  *   License: The package has a CRAN or OSI accepted license.
  *   Project management: Are the issue and PR trackers in a good shape, e.g. are there outstanding bugs, is it clear when feature requests are meant to be tackled?
