### 2025-04-24 #692 @maurolepore
@kelliejarcher thanks for following up. I re-announced the package among our editors. Once we find an available handling editor they will guide you through the rest of the process.

BTW, ideally, to keep this thread easier to follow it's best to try respond directly through the GitHub interface. Responding to the email notification duplicates the message you're responding to 😸
