### 2025-05-04 #692 @maurolepore
Here I mark the end of my EiC rotation and leave a short note for the next EiC.

* This is a stats submission.
* The package was announced twice but still lacks a handling editor.
