### 2025-06-02 #692 @kelliejarcher
I have been waiting to submit a peer-reviewed manuscript describing this R package to incorporate feedback from ROpenSci. After almost 3 months, I have not heard whether anyone has volunteered to review this package. Can you please let me know if there is something else I need to do or if I should simply forgo the ROpenSci review? I am new to this forum so am unclear on the process. Thanks for any information.
