### 2025-06-10 #692 @mpadge
> When I added this badge and updated github, the pgkcheck failed but it seems to be due to an R package that I am not explicitly using, as the error is ✖ Failed to build Rdsdp 1.0.6 (3.1s) Do you know how I can correct this? I am using R 4.5.0 and I checked that Rdsdp 1.0.6 is installed on my machine. Sorry for the naive question, I am new to pgkcheck and ROpenSci processes.

@kelliejarcher Those errors clearly come from the {Rdsdp} package itself, which is also generating lots of [current warnings on CRAN](https://cran.r-project.org/web/checks/check_results_Rdsdp.html). Since it's not even a direct dependency of your package, and does not have a public-facing git instance, you'll have to wait for a CRAN update for that to be fixed. (For reference: current version = 1.0.6).
