### 2025-06-30 #692 @lamtung16
> [@lamtung16](https://github.com/lamtung16) Thank you for your review of the hdcuremodels package. I have made your requested changes which are now on github. A few comments/questions:
>
> * I added your name as a reviewer in the DESCRIPTION file. Please verify I have your name and e-mail correctly identified.
> * For longer parameter descriptions in each help file, I changed to bulleted lists. Let me know if you would prefer the short lists (when there are 2 to 3 options) in any function also changed to a bulleted list.
> * I added a paragraph to the beginning of the vignette to more clearly define the notation being used.
> * In the vignette I explain what a and rho represent and what it means when they are 0 or rho is near 1.
> * I modified the generate_cure_data function to (a) not require the total number of covariates j to be divisible by the number of true relevant covariates n_true and (b) rename the columns to clearly identify which variables are the unpenalized covariates (prefix is U - the nonp default is 2) and which variables are penalized.
> * The plot function now has a label parameter that when TRUE, the variable labels are displayed in a legend for the trace plot.
>
> Let me know if you have further comments. Thanks again!

That’s a good improvement 👍😃   — thank you! I have a few comments:
- In the DESCRIPTION file, my email should be nguyenlamtung10@gmail.com, not lamtung16@gmail.com.
- In `generate_cure_data`, `nonp = 0`  does not work? Is it supposed to be not working?
- In `generate_cure_data`, the total number of features is `j + nonp`? Are `U1, U2, ...` noise?
- When I run `data <- generate_cure_data(n = 20, j = 3, n_true = 2)`, it doesn’t work. I’m not sure why.
- When I try the following code:
```
data <- generate_cure_data(n = 1000, j = 4, n_true = 2, nonp = 1)
fitem <- cureem(Surv(Time, Censor) ~ .,
  data = data$training,
  x_latency = data$training
)
plot(fitem, label = TRUE)
```

I'm not quite sure how to interpret this plot (can I say X1 and X2 are true features and X3 and X4 are noises?). It would be really helpful if you could provide a few simple example plots like this one, along with some explanation — for instance, something like:
"This dataset has 2 true features and 2 noise variables. As the penalty increases, the coefficients for the noise variables shrink to zero, while the true features remain. This shows the model's ability to distinguish signal from noise."
Having that kind of context would make the plot much more intuitive.

Honestly, I've been trying to make an example of a dataset having 3 features (2 true features and 1 noise). Then from the coefficients plot, when I have the big regularization parameter(s), I can see one coef shrinks to 0, the other two reach non-zero.

But overall, you guys have done a great job 👍
