### 2025-07-09 #692 @lamtung16
> Thanks again for your review of our package and helpful comments to improve the functions. The following changes have been made:
>
> * Your email address has been updated in the DESCRIPTION file.
> * The <generate_cure_data> function has been revised to allow <nonp = 0> and also to ensure it works for any combination of ,  and <n_true> and performs a check that  is greater than  <n_true>.
> * The help for <generate_cure_data> has been updated to better explain that covariates prefixed with "U" are unpenalized covariates (for example, suppose you want to coerce age and sex into the model but penalize all gene expression values). The vignette has also been expanded to fully explain all parameters of the <generate_cure_data> function. The "U" covariates are not associated with the outcome so it is correct that they are noise, but so are j - n_true of the penalized covariates.
> * Regarding your code example, the help file and vignette now better explain how to identify which penalized predictors are truly associated with the outcome. The colors and line types in the plot function were changed so that it is easier to identify the incidence ("I_") and latency ("L_") associations for each variable.
> * I added a few more generic methods to extract some interesting components from the resulting mixturecure object including dim, family, formula, logLik, and nobs.
> * This package passes devtools::check() but when uploading to GitHub it still gets the "Failed to build Rdsdp 1.0.6" but that seems to be due to a problem with Rdsdp 1.0.6 though I noticed https://cran.r-project.org/web/checks/check_results_Rdsdp.html indicates everything is OK with that package.

I checked, they look great 😃
