### 2025-07-09 #692 @lamtung16
> Thanks for your review, I really appreciate your comments that improved this package. Is the package ready to submit to CRAN and the manuscript to a journal?

I don't have the right to decide, it's up to the editor, but I think it's good for first manuscript submission 😺.
