### 2025-07-10 #692 @kelliejarcher
I doubt if any of these have reviewed for ROpenSci but here are a few names:

Panagiotis Papastamoulis ***@***.******@***.***>
Sharon Xiangwen Xie ***@***.******@***.***>
Yingwei Paul Peng ***@***.******@***.***>

Hope that helps,
Kellie

From: Toby Dylan Hocking ***@***.***>
Date: Wednesday, July 9, 2025 at 10:12 PM
To: ropensci/software-review ***@***.***>
Cc: Archer, Kellie ***@***.***>, Mention ***@***.***>
Subject: Re: [ropensci/software-review] hdcuremodels (Issue #692)
tdhock left a comment (ropensci/software-review#692) We are supposed to find two different reviewers, and I am still trying to find a second one. Please tell me if you have any suggestions of colleagues who you think may be willing and available

[Image removed by sender.]tdhock left a comment (ropensci/software-review#692)<https://urldefense.com/v3/__https:/github.com/ropensci/software-review/issues/692*issuecomment-3055024338__;Iw!!KGKeukY!wRDZWGrxu8bnW2ARIpoyoViCFCRFyiF4ikNmeFzZu3agsGLKVPkrVcPov-vcYr41WaaTRdGLjuJxNpbYOh3PBiDbhg$>

We are supposed to find two different reviewers, and I am still trying to find a second one. Please tell me if you have any suggestions of colleagues who you think may be willing and available to review.
