### 2025-09-16 #692 @ropensci-review-bot
Approved! Thanks @kelliejarcher for submitting and TBD for your reviews! :grin:
